其他文件/requirements.txt
argon2-cffi==20.1.0
asgiref==3.4.1
async-generator==1.10
attrs==21.2.0
backcall==0.2.0
backports.entry-points-selectable==1.1.0
beautifulsoup4==4.10.0
bleach==3.3.1
cachetools==4.2.4
certifi==2021.10.8
cffi==1.14.6
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
configparser==5.1.0
crayons==0.4.0
cryptography==35.0.0
debugpy==1.3.0
decorator==5.0.9
defusedxml==0.7.1
distlib==0.3.3
Django==3.2.8
django-extensions==3.1.5
django-filter==21.1
docopt==0.6.2
entrypoints==0.3
et-xmlfile==1.1.0
filelock==3.3.0
flake8==3.9.2
Flask==2.0.2
future==0.18.2
google-auth==2.3.3
google-auth-oauthlib==0.4.6
gspread==4.0.1
gunicorn==20.1.0
h11==0.12.0
httplib2==0.20.2
humanize==3.12.0
idna==3.3
ipykernel==6.0.3
ipython==7.25.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
liffpy==0.0.1a3
line-bot-sdk==1.8.0
lxml==4.6.3
MarkupSafe==2.0.1
matplotlib-inline==0.1.2
mccabe==0.6.1
mistune==0.8.4
mysql-connector==2.2.9
mysql-connector-python==8.0.27
mysqlclient==2.1.0
nbclient==0.5.3
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.0
numpy==1.21.1
oauth2client==4.1.3
oauthlib==3.1.1
openpyxl==3.0.7
outcome==1.1.0
packaging==21.0
pandas==1.3.0
pandocfilters==1.4.3
parso==0.8.2
pbr==5.8.1
pickleshare==0.7.5
platformdirs==2.4.0
prometheus-client==0.11.0
prompt-toolkit==3.0.19
protobuf==3.19.4
psycopg2==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
pyflakes==2.3.1
Pygments==2.9.0
PyMySQL==1.0.2
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
pythondialog==3.5.2
pytz==2021.1
pyvmomi==7.0.3
pyzmq==22.1.0
qtconsole==5.1.1
QtPy==1.9.0
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
selenium==4.0.0
Send2Trash==1.7.1
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.2.1
sqlparse==0.4.2
stevedore==3.5.0
tabulate==0.8.9
terminado==0.10.1
testpath==0.5.0
tornado==6.1
traitlets==5.0.5
trio==0.19.0
trio-websocket==0.9.2
urllib3==1.26.7
vconnector==0.6.0
virtualenv==20.8.1
virtualenv-clone==0.5.7
virtualenvwrapper==4.8.4
wcwidth==0.2.5
webdriver-manager==3.5.2
webencodings==0.5.1
Werkzeug==2.0.2
whitenoise==6.0.0
widgetsnbextension==3.5.1
wsproto==1.0.0
xlrd==2.0.1